Refactor out types from annotations module file
